# Prevent calls to uniquely_named_symbols() from iterating over zeroes of large sparse matrices