@oscarbenjamin 2024-01-11
> could someone please explain what
> sympy/matrices/expressions/trace.py : `_eval_rewrite_as_Sum()` does?

This would be called by:
```python
In [3]: t = Trace(Matrix([[1, 2], [3, 4]]))

In [4]: t
Out[4]: 
  ⎛⎡1  2⎤⎞
tr⎜⎢    ⎥⎟
  ⎝⎣3  4⎦⎠

In [5]: t.rewrite(Sum)
Out[5]: 5
```
It is not so much useful for an explicit matrix like this because we could use:
```python
In [7]: %time trace(eye(1000))
CPU times: user 141 ms, sys: 66.9 ms, total: 208 ms
Wall time: 206 ms
Out[7]: 1000

In [8]: %time eye(1000).trace()
CPU times: user 19.8 ms, sys: 21 µs, total: 19.9 ms
Wall time: 19.4 ms
Out[8]: 1000
```
The purpose of `.rewrite(Sum)` is for matrix expressions:
```python
In [15]: M = FunctionMatrix(n, n, Lambda((i, j), i**2+j**2))

In [16]: M
Out[16]: 
⎡ 2    2⎤
⎣i  + j ⎦

In [17]: Trace(M)
Out[17]: 
  ⎛⎡ 2    2⎤⎞
tr⎝⎣i  + j ⎦⎠

In [18]: Trace(M).rewrite(Sum)
Out[18]: 
   3         
2⋅n     2   n
──── - n  + ─
 3          3
```
It would make more sense if the rewrite didn't call doit though and returned the Sum unevaluated because `Trace.doit` already evaluates the summation:
```python
In [19]: Trace(M).doit()
Out[19]: 
   3         
2⋅n     2   n
──── - n  + ─
 3          3
```