@DeadSpheroid 2024-01-12
I wasnt entirely sure about changing it since i didnt know where the function was called, therefore i hadnt made that change and left the pr as draft. Thanks a lot though, should be good to go now!