@asmeurer 2024-01-12
This `uniquely_named_symbols` function is not public. Maybe it would be better to change the signature to `*exprs` to avoid this type of mistake. 

At any rate, I'm a little confused why it is used at all, instead of Dummy. Did we stop using Dummy for rewrites that generate bound symbols?