@smichr 2024-01-12
Except that they look bad when printed in text (having the leading underscore) they look ok when pretty printed and it would be ok to use Dummy here. The reason it isn't is because even though there is no internal problem there is an presentation issue if the variable `i` was already used in the expression being re-written as a `Sum` that uses `i` as its summation symbol:

![image](https://github.com/sympy/sympy/assets/90703/2889b9aa-9564-4be7-93e0-5aae9a9932b0)

The presentation makes it look like the result should be `6`, not `3i` (though the latter is correct). Hence the presentation issue, not an internal issue.