@oscarbenjamin 2024-01-12
It would still be better to remove the `doit` call in `Ttrace.rewrite(Sum)`.