@asmeurer 2024-01-12
We could use a uniquely named dummy symbol. The question is whether new bound symbols should be Dummy or not. I know there was an issue at some point about making this consistent, but I don't know if there was a consensus reached on it. `uniquely_named_symbol` just feels like a poor implementation of Dummy. For instance, it wouldn't be very "unique" if it is called twice on two different expressions.