@oscarbenjamin 2024-01-12
The primary reason to use Dummy is to avoid having a clash with symbols that a user might have defined but it usually does not matter whether bound symbols clash.