@asmeurer 2024-01-13
Also the leading underscore from print() is there exactly to avoid this ambiguity. Maybe the pretty printer should add the underscore or some other kind of marker when it detects another symbol with the same name in the expression. That way this "unique" name stuff is done at the printing level (as a side note, this same sort of thing could be done when an expression has two symbols with the same name but different assumptions). 

> The primary reason to use Dummy is to avoid having a clash with symbols that a user might have defined but it usually does not matter whether bound symbols clash.

Well, it does in the sense that `Sum(i, (i, 0, 3))` would be wrong if the summation `i` is supposed to be distinct from the existing free `i`. 

It just seems we should do things one way or another. Either a proper Dummy symbol, or a uniquely named normal symbol. Both have advantages and disadvantages, but it's the inconsistency that bothers me.